feat: working build cmd
